edge-db-mqtt/edge-db-mqtt/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
jsonschema
python-dotenv
requests
paho-mqtt
